LayoutTests/inspector/console/js-completions-expected.txt

Tests for the JavaScript code completion feature used by the console and other code inputs in various breakpoint editors.


== Running test suite: console.jsCompletions
-- Running test case: console.jsCompletions.variableCompletions
PASS: Completions should not contain myVariable by default.
PASS: Completions should not change before cache is cleared.
PASS: Completions should still not contain myVariable before cache is cleared.
PASS: Completions should change after cache is cleared.
PASS: Completions should contain myVariable after it's created by code run in console.
PASS: Completions should not change before cache is cleared.
PASS: Completions should still contain myVariable before cache is cleared.
PASS: Completions should change after cache is cleared.
PASS: Completions should not contain myVariable after it's deleted by code run in console.

-- Running test case: console.jsCompletions.completionOrdering
Completions with options={"baseString":"objectWithMixedPropertyKinds.","prefix":""} and experimentalShowCaseSensitiveAutocomplete=false:
[
  "aa",
  "Aa",
  "ab",
  "Ab",
  "__aa__",
  "__ab__"
]

Completions with options={"baseString":"objectWithMixedPropertyKinds.","prefix":"a"} and experimentalShowCaseSensitiveAutocomplete=false:
[
  "aa",
  "ab",
  "Aa",
  "Ab"
]

Completions with options={"baseString":"objectWithMixedPropertyKinds.","prefix":"A"} and experimentalShowCaseSensitiveAutocomplete=false:
[
  "Aa",
  "Ab",
  "aa",
  "ab"
]

Completions with options={"baseString":"objectWithMixedPropertyKinds.","prefix":"_"} and experimentalShowCaseSensitiveAutocomplete=false:
[
  "__aa__",
  "__ab__"
]

Completions with options={"baseString":"objectWithMixedPropertyKinds[","prefix":""} and experimentalShowCaseSensitiveAutocomplete=false:
[
  "1",
  "2",
  "10",
  "20",
  "\"__aa__\"",
  "\"__ab__\"",
  "\"aa\"",
  "\"Aa\"",
  "\"ab\"",
  "\"Ab\""
]

Completions with options={"baseString":"objectWithMixedPropertyKinds[","prefix":"1"} and experimentalShowCaseSensitiveAutocomplete=false:
[
  "1",
  "10"
]

Completions with options={"baseString":"objectWithMixedPropertyKinds[","prefix":"\"a"} and experimentalShowCaseSensitiveAutocomplete=false:
[
  "\"aa\"",
  "\"ab\"",
  "\"Aa\"",
  "\"Ab\""
]

Completions with options={"baseString":"objectWithMixedPropertyKinds[","prefix":"\"A"} and experimentalShowCaseSensitiveAutocomplete=false:
[
  "\"Aa\"",
  "\"Ab\"",
  "\"aa\"",
  "\"ab\""
]

Completions with options={"baseString":"objectWithMixedPropertyKinds[","prefix":"\"_"} and experimentalShowCaseSensitiveAutocomplete=false:
[
  "\"__aa__\"",
  "\"__ab__\""
]

Completions with options={"baseString":"objectWithMixedPropertyKinds.","prefix":""} and experimentalShowCaseSensitiveAutocomplete=true:
[
  "aa",
  "Aa",
  "ab",
  "Ab",
  "__aa__",
  "__ab__"
]

Completions with options={"baseString":"objectWithMixedPropertyKinds.","prefix":"a"} and experimentalShowCaseSensitiveAutocomplete=true:
[
  "aa",
  "ab"
]

Completions with options={"baseString":"objectWithMixedPropertyKinds.","prefix":"A"} and experimentalShowCaseSensitiveAutocomplete=true:
[
  "Aa",
  "Ab"
]

Completions with options={"baseString":"objectWithMixedPropertyKinds.","prefix":"_"} and experimentalShowCaseSensitiveAutocomplete=true:
[
  "__aa__",
  "__ab__"
]

Completions with options={"baseString":"objectWithMixedPropertyKinds[","prefix":""} and experimentalShowCaseSensitiveAutocomplete=true:
[
  "1",
  "2",
  "10",
  "20",
  "\"__aa__\"",
  "\"__ab__\"",
  "\"aa\"",
  "\"Aa\"",
  "\"ab\"",
  "\"Ab\""
]

Completions with options={"baseString":"objectWithMixedPropertyKinds[","prefix":"1"} and experimentalShowCaseSensitiveAutocomplete=true:
[
  "1",
  "10"
]

Completions with options={"baseString":"objectWithMixedPropertyKinds[","prefix":"\"a"} and experimentalShowCaseSensitiveAutocomplete=true:
[
  "\"aa\"",
  "\"ab\""
]

Completions with options={"baseString":"objectWithMixedPropertyKinds[","prefix":"\"A"} and experimentalShowCaseSensitiveAutocomplete=true:
[
  "\"Aa\"",
  "\"Ab\""
]

Completions with options={"baseString":"objectWithMixedPropertyKinds[","prefix":"\"_"} and experimentalShowCaseSensitiveAutocomplete=true:
[
  "\"__aa__\"",
  "\"__ab__\""
]
